fix: header anchors and auth pages
