+++ vite-project/src/pages/Register.tsx
@@ -1,0 +1,16 @@
+import AuthHeader from "../components/Authentication/AuthHeader"
+import RegisterComponent from "../components/Authentication/RegisterComponent"
+import { AuthPageStyle } from "../components/Authentication/style"
+
+function Register(){
+    return(<>
+
+    <AuthHeader/>
+    <AuthPageStyle>
+       <RegisterComponent/>
+    </AuthPageStyle>
+    
+    </>)
+}
+
+export default Register
