+++ vite-project/src/routes/index.tsx
@@ -1,13 +1,15 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Homepage from '../pages/Homepage';
 import Login from '../pages/Login';
+import Register from '../pages/Register';
 
 function App() {
   return (
     <Router>
       <Routes>
         <Route path="/" element={<Homepage/>} />
-        <Route path="/auth" element={<Login/>} />
+        <Route path="/auth/register" element={<Register/>} />
+        <Route path="/auth/login" element={<Login/>} />
       </Routes>
     </Router>
   );
